[Crowcroft 1995] updated to reflect the edition the Durham Main Library actually has.
Half an hour lost, and no real effect.  Except to see that the amount of
printed journals the IEEE publishes is vast.  And finding them is not easy.
Does anybody actually read this stuff?  And yeah, the quality varies.  IEEE
Network is one of the lower-quality journals, or so it seems on the surface at
least.