One more run of latex.  Now it's three runs.  Where will this end?!
Reason:  The table reference is numbered wrongly.  This still does not fix it.

More to the point, I would really like to know what the proper invocation of
all these partial commands is supposed to be.  I can remember clearly pdflatex
or somesuch does all this on its own, like a good program should.  That would
be nice.